+++ docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 2025-02-25
+
+- Ensure that the registration file is updated when the configuration changes.
+
 ### 2025-02-13
 
 - Add changelog for tracking user-relevant changes.